new: swap client and server architecture
in reality, the rover should be the server
this code is lightly tested however should be fine
