+++ README.md
@@ -6,13 +6,12 @@ Programming Tree for FIU Lunabotics Rover
 
 Make sure you have:
 
-- Linux is required, the controller code uses Linux's native [evdev interface](https://en.wikipedia.org/wiki/Evdev).
+- Linux, as the controller code uses Linux's native [evdev interface](https://en.wikipedia.org/wiki/Evdev).
 - Either `pip` or manually installed the requirements from your distro.
 
 ### Pip
 
-- Install `python3-pip` (or equiv.), if too old, do
-  `pip3 install --upgrade pip` after.
+- Install `python3-pip` (or equivalent)
 
 Then run:
 
@@ -22,16 +21,16 @@ pip install --user -r requirements.txt
 
 ## Running
 
-On server (PC which has controller connected to it.) run:
-
-```bash
-./controller/server.py --public
-```
-
-On client (Nano) run:
+On client (PC which has controller connected to it.) run:
 
 ```bash
 ./controller/client.py [SERVER_IP]
+```
+
+On server (rover/Raspberry Pi) run:
+
+```bash
+./controller/server.py --public
 ```
 
 To see all available options, both scripts support `--help`.
